requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==21.2.0
    # via python-rsync-backup (setup.py)
bcrypt==3.2.0
    # via paramiko
cffi==1.14.5
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
cryptography==3.4.7
    # via paramiko
iso8601==0.1.14
    # via python-rsync-backup (setup.py)
paramiko==2.7.2
    # via spur
pycparser==2.20
    # via cffi
pynacl==1.4.0
    # via paramiko
six==1.16.0
    # via
    #   bcrypt
    #   pynacl
spur==0.3.22
    # via python-rsync-backup (setup.py)
toml==0.10.2
    # via python-rsync-backup (setup.py)
xdg==5.0.2
    # via python-rsync-backup (setup.py)
